+++ source/exe/BUILD
@@ -41,7 +41,6 @@ envoy_cc_library(
         "//source/common/stats:thread_local_store_lib",
         "//source/server:drain_manager_lib",
         "//source/server:listener_hooks_lib",
-        "//source/server:options_lib",
         "//source/server:server_base_lib",
     ] + select({
         "//bazel:windows_x86_64": envoy_all_extensions(WINDOWS_SKIP_TARGETS),
@@ -85,7 +84,7 @@ envoy_cc_library(
         "//source/server:hot_restart_lib",
         "//source/server:hot_restart_nop_lib",
         "//source/server:listener_hooks_lib",
-        "//source/server:options_lib",
+        "//source/server:options_base",
         "//source/server:server_base_lib",
     ] + envoy_select_signal_trace([
         "//source/common/signal:sigaction_lib",
@@ -97,22 +96,22 @@ envoy_cc_library(
     name = "main_common_lib",
     srcs = [
         "main_common.cc",
-        "stripped_main_base.cc",
     ],
     hdrs = [
         "main_common.h",
-        "stripped_main_base.h",
     ],
     deps = [
         ":envoy_common_lib",
         ":platform_impl_lib",
         ":process_wide_lib",
+        ":stripped_main_base_lib",
         "//source/common/api:os_sys_calls_lib",
         "//source/common/common:compiler_requirements_lib",
         "//source/common/common:perf_annotation_lib",
         "//source/common/grpc:google_grpc_context_lib",
         "//source/server:hot_restart_lib",
         "//source/server:hot_restart_nop_lib",
+        "//source/server:options_lib",
         "//source/server/config_validation:server_lib",
     ] + envoy_select_signal_trace([
         "//source/common/signal:sigaction_lib",
@@ -144,15 +143,8 @@ envoy_cc_library(
 
 envoy_cc_library(
     name = "envoy_main_common_with_core_extensions_lib",
-    srcs = [
-        "main_common.cc",
-        "stripped_main_base.cc",
-    ],
-    hdrs = [
-        "main_common.h",
-        "stripped_main_base.h",
-    ],
     deps = [
+               ":main_common_lib",
                ":platform_impl_lib",
                ":process_wide_lib",
                "//envoy/server:platform_interface",
