+++ source/server/options_impl.cc
@@ -33,10 +33,6 @@ std::vector<std::string> toArgsVector(int argc, const char* const* argv) {
   return args;
 }
 
-void throwMalformedArgExceptionOrPanic(std::string message) {
-  throwExceptionOrPanic(MalformedArgvException, message);
-}
-
 } // namespace
 
 OptionsImpl::OptionsImpl(int argc, const char* const* argv,
@@ -194,7 +190,11 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
   cpuset_threads_ = cpuset_threads.getValue();
 
   if (log_level.isSet()) {
-    log_level_ = parseAndValidateLogLevel(log_level.getValue());
+    auto status_or_error = parseAndValidateLogLevel(log_level.getValue());
+    if (!status_or_error.status().ok()) {
+      logError(std::string(status_or_error.status().message()));
+    }
+    log_level_ = status_or_error.value();
   } else {
     log_level_ = default_log_level;
   }
@@ -214,7 +214,7 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
     mode_ = Server::Mode::InitOnly;
   } else {
     const std::string message = fmt::format("error: unknown mode '{}'", mode.getValue());
-    throwMalformedArgExceptionOrPanic(message);
+    throw MalformedArgvException(message);
   }
 
   if (local_address_ip_version.getValue() == "v4") {
@@ -224,7 +224,7 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
   } else {
     const std::string message =
         fmt::format("error: unknown IP address version '{}'", local_address_ip_version.getValue());
-    throwMalformedArgExceptionOrPanic(message);
+    throw MalformedArgvException(message);
   }
   base_id_ = base_id.getValue();
   use_dynamic_base_id_ = use_dynamic_base_id.getValue();
@@ -234,7 +234,7 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
   if (use_dynamic_base_id_ && restart_epoch_ > 0) {
     const std::string message = fmt::format(
         "error: cannot use --restart-epoch={} with --use-dynamic-base-id", restart_epoch_);
-    throwMalformedArgExceptionOrPanic(message);
+    throw MalformedArgvException(message);
   }
 
   if (!concurrency.isSet() && cpuset_threads_) {
@@ -275,8 +275,8 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
   } else {
     uint64_t socket_mode_helper;
     if (!StringUtil::atoull(socket_mode.getValue().c_str(), socket_mode_helper, 8)) {
-      throwExceptionOrPanic(MalformedArgvException,
-                            fmt::format("error: invalid socket-mode '{}'", socket_mode.getValue()));
+      throw MalformedArgvException(
+          fmt::format("error: invalid socket-mode '{}'", socket_mode.getValue()));
     }
     socket_mode_ = socket_mode_helper;
   }
@@ -286,13 +286,13 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
   } else if (drain_strategy.getValue() == "gradual") {
     drain_strategy_ = Server::DrainStrategy::Gradual;
   } else {
-    throwExceptionOrPanic(MalformedArgvException,
-                          fmt::format("error: unknown drain-strategy '{}'", mode.getValue()));
+    throw MalformedArgvException(
+        fmt::format("error: unknown drain-strategy '{}'", mode.getValue()));
   }
 
   if (hot_restart_version_option.getValue()) {
     std::cerr << hot_restart_version_cb(!hot_restart_disabled_);
-    throwExceptionOrPanic(NoServingException, "NoServingException");
+    throw NoServingException("NoServingException");
   }
 
   if (!disable_extensions.getValue().empty()) {
@@ -305,22 +305,20 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
       std::vector<absl::string_view> cli_tag_pair_tokens =
           absl::StrSplit(cli_tag_pair, absl::MaxSplits(':', 1));
       if (cli_tag_pair_tokens.size() != 2) {
-        throwExceptionOrPanic(MalformedArgvException,
-                              fmt::format("error: misformatted stats-tag '{}'", cli_tag_pair));
+        throw MalformedArgvException(
+            fmt::format("error: misformatted stats-tag '{}'", cli_tag_pair));
       }
 
       auto name = cli_tag_pair_tokens[0];
       if (!Stats::TagUtility::isTagNameValid(name)) {
-        throwExceptionOrPanic(
-            MalformedArgvException,
+        throw MalformedArgvException(
             fmt::format("error: misformatted stats-tag '{}' contains invalid char in '{}'",
                         cli_tag_pair, name));
       }
 
       auto value = cli_tag_pair_tokens[1];
       if (!Stats::TagUtility::isTagValueValid(value)) {
-        throwExceptionOrPanic(
-            MalformedArgvException,
+        throw MalformedArgvException(
             fmt::format("error: misformatted stats-tag '{}' contains invalid char in '{}'",
                         cli_tag_pair, value));
       }
@@ -328,26 +326,6 @@ OptionsImpl::OptionsImpl(std::vector<std::string> args,
       stats_tags_.emplace_back(Stats::Tag{std::string(name), std::string(value)});
     }
   }
-}
-
-spdlog::level::level_enum OptionsImpl::parseAndValidateLogLevel(absl::string_view log_level) {
-  if (log_level == "warn") {
-    return spdlog::level::level_enum::warn;
-  }
-
-  size_t level_to_use = std::numeric_limits<size_t>::max();
-  for (size_t i = 0; i < ARRAY_SIZE(spdlog::level::level_string_views); i++) {
-    spdlog::string_view_t spd_log_level = spdlog::level::level_string_views[i];
-    if (log_level == absl::string_view(spd_log_level.data(), spd_log_level.size())) {
-      level_to_use = i;
-      break;
-    }
-  }
-
-  if (level_to_use == std::numeric_limits<size_t>::max()) {
-    logError(fmt::format("error: invalid log level specified '{}'", log_level));
-  }
-  return static_cast<spdlog::level::level_enum>(level_to_use);
 }
 
 std::string OptionsImpl::allowedLogLevels() {
@@ -374,7 +352,11 @@ void OptionsImpl::parseComponentLogLevels(const std::string& component_log_level
       logError(fmt::format("error: component log level not correctly specified '{}'", level));
     }
     std::string log_name = log_name_level[0];
-    spdlog::level::level_enum log_level = parseAndValidateLogLevel(log_name_level[1]);
+    auto status_or_error = parseAndValidateLogLevel(log_name_level[1]);
+    if (!status_or_error.status().ok()) {
+      logError(std::string(status_or_error.status().message()));
+    }
+    spdlog::level::level_enum log_level = status_or_error.value();
     Logger::Logger* logger_to_change = Logger::Registry::logger(log_name);
     if (!logger_to_change) {
       logError(fmt::format("error: invalid component specified '{}'", log_name));
@@ -383,9 +365,7 @@ void OptionsImpl::parseComponentLogLevels(const std::string& component_log_level
   }
 }
 
-uint32_t OptionsImpl::count() const { return count_; }
-
-void OptionsImpl::logError(const std::string& error) { throwMalformedArgExceptionOrPanic(error); }
+void OptionsImpl::logError(const std::string& error) { throw MalformedArgvException(error); }
 
 Server::CommandLineOptionsPtr OptionsImpl::toCommandLineOptions() const {
   Server::CommandLineOptionsPtr command_line_options =
@@ -449,25 +429,11 @@ Server::CommandLineOptionsPtr OptionsImpl::toCommandLineOptions() const {
 }
 
 OptionsImpl::OptionsImpl(const std::string& service_cluster, const std::string& service_node,
-                         const std::string& service_zone, spdlog::level::level_enum log_level)
-    : log_level_(log_level), service_cluster_(service_cluster), service_node_(service_node),
-      service_zone_(service_zone) {}
-
-void OptionsImpl::disableExtensions(const std::vector<std::string>& names) {
-  for (const auto& name : names) {
-    const std::vector<absl::string_view> parts = absl::StrSplit(name, absl::MaxSplits('/', 1));
-
-    if (parts.size() != 2) {
-      ENVOY_LOG_MISC(warn, "failed to disable invalid extension name '{}'", name);
-      continue;
-    }
-
-    if (Registry::FactoryCategoryRegistry::disableFactory(parts[0], parts[1])) {
-      ENVOY_LOG_MISC(info, "disabled extension '{}'", name);
-    } else {
-      ENVOY_LOG_MISC(warn, "failed to disable unknown extension '{}'", name);
-    }
-  }
+                         const std::string& service_zone, spdlog::level::level_enum log_level) {
+  setLogLevel(log_level);
+  setServiceClusterName(service_cluster);
+  setServiceNodeName(service_node);
+  setServiceZone(service_zone);
 }
 
 } // namespace Envoy
